Added Custom GCS function
Looks for Custom-file/gcs-custom-search.php file, reads the code and
cuts out the php. Adds a new database row called gcs_address and adds
the code to it. Deletes the folder. Adds a gcs settings page on the
admin page that allows to add or change gcs code with a form.